% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook.R
\docType{data}
\name{fb_bot}
\alias{fb_bot}
\title{Facebook bot}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
fb_bot
}
\description{
Class with function to extract data from facebook posts using webdriver and phantomjs
}
\section{Usage}{


\preformatted{

bot_facebook = fb_bot$new()
bot_facebook$login(username = <facebook username>, password = <facebook password>)
posts = bot_facebook$get_posts(pagename = "ameliarueda", n = 50, reactions = T)
}
}

\section{Details}{

\code{fb_bot$new()} Creates a new fbbot.

\code{bot_facebook$login()} Log in to your facebook account.

\code{bot_facebook$get_posts} Log in to your facebook account.
}

\keyword{datasets}
