% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium_server.R
\name{find_emails}
\alias{find_emails}
\title{Extract email}
\usage{
find_emails(text = NA_character_,
  email_services = c("gmail|yahoo|outlook|hotmail|live"))
}
\arguments{
\item{text}{string}

\item{email_services}{email services}
}
\description{
Extract email
}
