% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook.R
\name{get_comments}
\alias{get_comments}
\title{Get Facebook comments}
\usage{
get_comments(x = NULL, page_id = NA_character_, post_id = NA_character_)
}
\arguments{
\item{x}{scrapeR_session objetc}

\item{page_id}{facebook page id}

\item{post_id}{facebook post id}
}
\description{
Get Facebook comments
}
