% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook.R
\name{get_fb_posts}
\alias{get_fb_posts}
\title{Get Facebook post}
\usage{
get_fb_posts(x = NULL, pagename = NA_character_, n = 10L,
  reactions = FALSE, commets = FALSE, shares = FALSE)
}
\arguments{
\item{x}{session}

\item{pagename}{page name}

\item{n}{n of posts}

\item{reactions}{if TRUE add reactions}

\item{commets}{if TRUE add comments}

\item{shares}{if TRUE add shares}
}
\description{
Get Facebook post
}
