% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook.R
\name{get_page_info}
\alias{get_page_info}
\title{get facebook page info}
\usage{
get_page_info(x = NULL, page_id = NA_character_)
}
\arguments{
\item{x}{sesson}

\item{page_id}{page id}
}
\description{
get facebook page info
}
