% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook.R
\name{get_reactions}
\alias{get_reactions}
\title{Get reactions}
\usage{
get_reactions(x = NULL, page_id = NA_character_, post_id = NA_character_,
  silence = T)
}
\arguments{
\item{x}{scrapeR_session}

\item{page_id}{page id}

\item{post_id}{post id}

\item{silence}{otro}
}
\description{
Get reactions
}
