% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook.R
\name{get_reactions_}
\alias{get_reactions_}
\title{get_reactions_}
\usage{
get_reactions_(x = NULL, page_id = NA_character_,
  post_id = NA_character_, silence = T)
}
\description{
get_reactions_
}
\keyword{internal}
