% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook.R
\name{get_reactions_by_user}
\alias{get_reactions_by_user}
\title{Get reactions by user}
\usage{
get_reactions_by_user(x = NULL, post_id = NA_character_)
}
\arguments{
\item{x}{scrapeR_session}

\item{post_id}{facebook post id}
}
\description{
Get reactions by user
}
