% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook.R
\name{get_reactions_by_user_}
\alias{get_reactions_by_user_}
\title{get_reactions_by_user_}
\usage{
get_reactions_by_user_(x = NULL, post_id = NA_character_)
}
\description{
get_reactions_by_user_
}
\keyword{internal}
