% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook.R
\name{get_shares}
\alias{get_shares}
\title{Get shares}
\usage{
get_shares(x = NULL, page_id = NA_character_,
  post_id = NA_character_)
}
\arguments{
\item{x}{session}

\item{page_id}{page id}

\item{post_id}{post id}
}
\description{
Get shares
}
