% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{get_tw_posts}
\alias{get_tw_posts}
\title{Get Twitter posts}
\usage{
get_tw_posts(x, page_id = NA_character_, n = NA_integer_)
}
\arguments{
\item{x}{session}

\item{page_id}{page_id}

\item{n}{n of posts}
}
\description{
Get Twitter posts
}
