% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook.R
\name{get_user_info}
\alias{get_user_info}
\title{Get public user info}
\usage{
get_user_info(x = NULL, username = NA_character_)
}
\arguments{
\item{x}{session}

\item{username}{username}
}
\description{
Get public user info
}
