% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phantomjs.R
\name{install_phantomjs}
\alias{install_phantomjs}
\title{Install PhantomJS}
\usage{
install_phantomjs(
  version = "2.1.1",
  baseURL = "https://github.com/wch/webshot/releases/download/v0.3.1/"
)
}
\arguments{
\item{version}{The version number of PhantomJS}

\item{baseURL}{The base URL for the location of PhantomJS binaries for download.}
}
\description{
Install PhantomJS
}
