% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium_server.R
\name{load_more}
\alias{load_more}
\title{load more items}
\usage{
load_more(x, xpath = NA_character_)
}
\arguments{
\item{x}{scrapeR_session}

\item{xpath}{xpath expresion}
}
\description{
load more items
}
