% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook.R
\name{login_facebook}
\alias{login_facebook}
\title{Login in facebook}
\usage{
login_facebook(x = NULL, username = NA_character_,
  password = NA_character_)
}
\arguments{
\item{x}{scrapeR_session}

\item{username}{facebook email}

\item{password}{facebook password}
}
\description{
Login in facebook
}
