% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{login_twitter}
\alias{login_twitter}
\title{Login on Twitter}
\usage{
login_twitter(x, username = NA_character_, password = NA_character_)
}
\arguments{
\item{x}{session}

\item{username}{username}

\item{password}{password}
}
\description{
Login on Twitter
}
