% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium_server.R
\name{scroll_n}
\alias{scroll_n}
\title{Load n items}
\usage{
scroll_n(x = NULL, xpath = NA_character_, n = NA_integer_)
}
\arguments{
\item{x}{scrapeR_session}

\item{xpath}{xpath expresion}

\item{n}{n}
}
\description{
Load n items
}
