% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialscrapeR.R
\docType{package}
\name{socialscrapeR}
\alias{socialscrapeR}
\title{Extraction and cleaning of data from the web}
\description{
Extraction and cleaning of data from the web
}
\author{
Oldemar Rodriguez Rojas \cr
Maintainer: Carlos Aguero B <carlos.aguero@promidat.com>
}
