% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium_server.R
\name{start_server}
\alias{start_server}
\title{Start selenium server}
\usage{
start_server(port = 4567L, headless = T, verbose = FALSE,
  settings.images = 2L, chromever. = NA)
}
\arguments{
\item{port}{Port to run on}

\item{headless}{headless}

\item{verbose}{If TRUE, include status messages (if any)}

\item{settings.images}{settings.images}
}
\description{
Start selenium server
}
