% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{tw_bot}
\alias{tw_bot}
\title{Twitter bot}
\description{
Class with function to extract data from twitter posts using webdriver and phantomjs
}
\section{Usage}{


\preformatted{

bot_twitter = tw_bot$new()
bot_twitter$login(username = <facebook username>, password = <facebook password>)
posts = bot_twitter$get_posts(pagename = "OVSICORI_UNA", n = 100)
}
}

\section{Details}{

\code{fb_bot$new()} Creates a new fbbot.

\code{bot_facebook$login()} Log in to your facebook account.

\code{bot_facebook$get_posts} Log in to your facebook account.
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{tw_bot$new()}}
\item \href{#method-print}{\code{tw_bot$print()}}
\item \href{#method-login}{\code{tw_bot$login()}}
\item \href{#method-get_posts}{\code{tw_bot$get_posts()}}
\item \href{#method-get_shot}{\code{tw_bot$get_shot()}}
\item \href{#method-clone}{\code{tw_bot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tw_bot$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tw_bot$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-login"></a>}}
\subsection{Method \code{login()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tw_bot$login(username = NA_character_, password = NA_character_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_posts"></a>}}
\subsection{Method \code{get_posts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tw_bot$get_posts(pagename = NA_character_, n = NA_integer_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_shot"></a>}}
\subsection{Method \code{get_shot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tw_bot$get_shot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tw_bot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
