% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook.R
\name{fb_bot}
\alias{fb_bot}
\title{Facebook bot}
\description{
Class with function to extract data from facebook posts using webdriver and phantomjs
}
\section{Usage}{


\preformatted{

bot_facebook = fb_bot$new()
bot_facebook$login(username = <facebook username>, password = <facebook password>)
posts = bot_facebook$get_posts(pagename = "ameliarueda", n = 50, reactions = T)
}
}

\section{Details}{

\code{fb_bot$new()} Creates a new fbbot.

\code{bot_facebook$login()} Log in to your facebook account.

\code{bot_facebook$get_posts} Log in to your facebook account.
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{fb_bot$new()}}
\item \href{#method-print}{\code{fb_bot$print()}}
\item \href{#method-display}{\code{fb_bot$display()}}
\item \href{#method-login}{\code{fb_bot$login()}}
\item \href{#method-get_posts}{\code{fb_bot$get_posts()}}
\item \href{#method-get_reactions}{\code{fb_bot$get_reactions()}}
\item \href{#method-get_comments}{\code{fb_bot$get_comments()}}
\item \href{#method-find_emails}{\code{fb_bot$find_emails()}}
\item \href{#method-clone}{\code{fb_bot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fb_bot$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fb_bot$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-display"></a>}}
\subsection{Method \code{display()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fb_bot$display()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-login"></a>}}
\subsection{Method \code{login()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fb_bot$login(username = NA_character_, password = NA_character_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_posts"></a>}}
\subsection{Method \code{get_posts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fb_bot$get_posts(pagename = NA_character_, n = NA_integer_, reactions = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_reactions"></a>}}
\subsection{Method \code{get_reactions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fb_bot$get_reactions(post_id = NA_character_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_comments"></a>}}
\subsection{Method \code{get_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fb_bot$get_comments(page_id = NA_character_, post_id = NA_character_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-find_emails"></a>}}
\subsection{Method \code{find_emails()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fb_bot$find_emails(
  text = NA_character_,
  email_services = c("gmail|yahoo|outlook|hotmail|live")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fb_bot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
