% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\docType{data}
\name{tw_bot}
\alias{tw_bot}
\title{Twitter bot}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
tw_bot
}
\description{
Class with function to extract data from twitter posts using webdriver and phantomjs
}
\section{Usage}{


\preformatted{

bot_twitter = tw_bot$new()
bot_twitter$login(username = <facebook username>, password = <facebook password>)
posts = bot_twitter$get_posts(pagename = "OVSICORI_UNA", n = 100)
}
}

\section{Details}{

\code{fb_bot$new()} Creates a new fbbot.

\code{bot_facebook$login()} Log in to your facebook account.

\code{bot_facebook$get_posts} Log in to your facebook account.
}

\keyword{datasets}
